apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.2.1

psycopg[binary]==3.2.1
SQLAlchemy==2.0.36
alembic==1.13.3

redis==5.0.8
boto3==1.35.89
botocore==1.35.89
python-multipart==0.0.9

PyJWT==2.10.1
argon2-cffi==23.1.0
email-validator==2.2.0
pytest==8.3.5
httpx==0.27.2
